Misc: change to logging utils only
